Setup for generating render helpers
