+++ examples/README.md
@@ -1,0 +1,1 @@
+The "SunnyLand" a texture pack is by Ansimuz (https://ansimuz.itch.io/sunny-land-pixel-game-art)
