+++ examples/minified_all_features.json
@@ -1,0 +1,1 @@
+{"__header__":{"fileType":"LEd Project JSON","app":"LEd","appAuthor":"Sebastien Benard","appVersion":"0.2.1-beta","url":"https://deepnight.net/tools/led-2d-level-editor/"},"name":"simple_with_array","jsonVersion":"0.2.1","defaultPivotX":0.5,"defaultPivotY":0.5,"defaultGridSize":16,"bgColor":"#27272A","nextUid":10,"minifyJson":true,"defs":{"layers":[{"identifier":"IntGrid","type":"IntGrid","uid":7,"gridSize":16,"displayOpacity":1.0,"intGridValues":[{"identifier":"dirt","color":"#E61919"}],"autoTilesetDefUid":2,"autoRuleGroups":[{"uid":8,"name":"New group","active":true,"collapsed":false,"rules":[{"uid":9,"active":true,"size":3,"tileIds":[0],"chance":1.0,"pattern":[0,1,0,0,0,0,0,0,0],"flipX":false,"flipY":false,"perlinActive":false,"perlinSeed":4908307,"perlinScale":0.2,"perlinOctaves":2}]}],"autoSourceLayerDefUid":null,"tilesetDefUid":null,"tilePivotX":0,"tilePivotY":0},{"identifier":"Entities","type":"Entities","uid":4,"gridSize":16,"displayOpacity":1.0,"intGridValues":[{"identifier":null,"color":"#000000"}],"autoTilesetDefUid":null,"autoRuleGroups":[],"autoSourceLayerDefUid":null,"tilesetDefUid":null,"tilePivotX":0,"tilePivotY":0},{"identifier":"Tiles","type":"Tiles","uid":1,"gridSize":16,"displayOpacity":1.0,"intGridValues":[{"identifier":null,"color":"#000000"}],"autoTilesetDefUid":null,"autoRuleGroups":[],"autoSourceLayerDefUid":null,"tilesetDefUid":2,"tilePivotX":0,"tilePivotY":0}],"entities":[{"identifier":"Test_entity","uid":3,"width":16,"height":16,"color":"#94D9B3","renderMode":"Cross","tilesetId":null,"tileId":null,"tileRenderMode":"Stretch","maxPerLevel":1,"limitBehavior":"DiscardOldOnes","pivotX":0.5,"pivotY":0.5,"fieldDefs":[{"identifier":"float_array_test","__type":"Array<Float>","uid":5,"type":"F_Float","isArray":true,"canBeNull":false,"arrayMinLength":null,"arrayMaxLength":null,"editorDisplayMode":"NameAndValue","editorDisplayPos":"Above","editorAlwaysShow":false,"min":null,"max":null,"defaultOverride":null}]}],"tilesets":[{"identifier":"SunnyLand_by_Ansimuz","uid":2,"relPath":"SunnyLand_by_Ansimuz.png","pxWid":400,"pxHei":368,"tileGridSize":16,"spacing":0,"padding":16,"savedSelections":[]}],"enums":[],"externalEnums":[]},"levels":[{"identifier":"Test_level","uid":0,"pxWid":48,"pxHei":48,"layerInstances":[{"__identifier":"IntGrid","__type":"IntGrid","__cWid":3,"__cHei":3,"__gridSize":16,"levelId":0,"layerDefUid":7,"pxOffsetX":0,"pxOffsetY":0,"intGrid":[{"coordId":0,"v":0},{"coordId":1,"v":0},{"coordId":2,"v":0}],"autoTiles":[{"ruleId":9,"tiles":[{"coordId":3,"tileId":0,"flips":0,"__tileX":16,"__tileY":16},{"coordId":4,"tileId":0,"flips":0,"__tileX":16,"__tileY":16},{"coordId":5,"tileId":0,"flips":0,"__tileX":16,"__tileY":16}]}],"seed":8928229,"gridTiles":[],"entityInstances":[]},{"__identifier":"Entities","__type":"Entities","__cWid":3,"__cHei":3,"__gridSize":16,"levelId":0,"layerDefUid":4,"pxOffsetX":0,"pxOffsetY":0,"intGrid":[],"autoTiles":[],"seed":8212170,"gridTiles":[],"entityInstances":[{"__identifier":"Test_entity","__cx":2,"__cy":0,"defUid":3,"x":40,"y":8,"fieldInstances":[{"__identifier":"float_array_test","__value":[],"__type":"Array<Float>","defUid":5,"realEditorValues":[]}]}]},{"__identifier":"Tiles","__type":"Tiles","__cWid":3,"__cHei":3,"__gridSize":16,"levelId":0,"layerDefUid":1,"pxOffsetX":0,"pxOffsetY":0,"intGrid":[],"autoTiles":[],"seed":598312,"gridTiles":[{"coordId":5,"tileId":0,"__tileX":16,"__tileY":16},{"coordId":6,"tileId":46,"__tileX":16,"__tileY":48},{"coordId":7,"tileId":46,"__tileX":16,"__tileY":48},{"coordId":8,"tileId":46,"__tileX":16,"__tileY":48}],"entityInstances":[]}]}]}
